# Conflicts:
#	src/main/java/com/chicmic/trainingModule/Controller/AssignTaskController/AssignTaskCRUD.java
#	src/main/java/com/chicmic/trainingModule/Controller/PlanController/TraineePlanCRUD.java
#	src/main/java/com/chicmic/trainingModule/Service/AssignTaskService/AssignTaskResponseMapper.java
#	src/main/java/com/chicmic/trainingModule/Service/DashboardService/DashboardService.java
#	src/main/resources/application.properties

+++ src/main/java/com/chicmic/trainingModule/Dto/DashboardDto/FeedbackResponseDto.java
@@ -9,6 +9,6 @@ import java.util.Date;
 public class FeedbackResponseDto {
     private String name;
     private String date;
-    private Float rating;
+    private Double rating;
     private String feedback;
 }
